requirements.txt
fastapi
uvicorn
python-multipart
pyatv
torch
open_clip_torch
pillow
opencv-python
numpy